+++ kohya_gui/class_basic_training.py
