+++ kohya_gui/class_command_executor.py
@@ -1,6 +1,6 @@
 import subprocess
 import psutil
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
