+++ kohya_gui/class_sample_images.py
@@ -3,7 +3,7 @@ import os
 import gradio as gr
 from easygui import msgbox
 
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
