+++ kohya_gui/class_sdxl_parameters.py
