+++ kohya_gui/convert_model_gui.py
@@ -5,7 +5,7 @@ import os
 import shutil
 from .common_gui import get_folder_path, get_file_path
 
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
