+++ kohya_gui/custom_logging.py
