+++ kohya_gui/dreambooth_folder_creation_gui.py
@@ -4,7 +4,7 @@ from .common_gui import get_folder_path
 import shutil
 import os
 
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
