+++ kohya_gui/localization.py
