+++ kohya_gui/manual_caption_gui.py
@@ -5,7 +5,7 @@ from math import ceil
 import os
 import re
 
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
