+++ kohya_gui/resize_lora_gui.py
@@ -4,7 +4,7 @@ import subprocess
 import os
 from .common_gui import get_saveasfilename_path, get_file_path
 
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
