+++ kohya_gui/svd_merge_lora_gui.py
@@ -8,7 +8,7 @@ from .common_gui import (
     get_file_path,
 )
 
-from library.custom_logging import setup_logging
+from .custom_logging import setup_logging
 
 # Set up logging
 log = setup_logging()
