+++ kohya_gui/utilities.py
@@ -6,13 +6,13 @@
 import gradio as gr
 import os
 import argparse
-from library.basic_caption_gui import gradio_basic_caption_gui_tab
-from library.convert_model_gui import gradio_convert_model_tab
-from library.blip_caption_gui import gradio_blip_caption_gui_tab
-from library.git_caption_gui import gradio_git_caption_gui_tab
-from library.wd14_caption_gui import gradio_wd14_caption_gui_tab
-from library.manual_caption_gui import gradio_manual_caption_gui_tab
-from library.group_images_gui import gradio_group_images_gui_tab
+from .basic_caption_gui import gradio_basic_caption_gui_tab
+from .convert_model_gui import gradio_convert_model_tab
+from .blip_caption_gui import gradio_blip_caption_gui_tab
+from .git_caption_gui import gradio_git_caption_gui_tab
+from .wd14_caption_gui import gradio_wd14_caption_gui_tab
+from .manual_caption_gui import gradio_manual_caption_gui_tab
+from .group_images_gui import gradio_group_images_gui_tab
 
 
 def utilities_tab(
