+++ textual_inversion_gui.py
@@ -11,7 +11,7 @@ import subprocess
 import pathlib
 import argparse
 from datetime import datetime
-from library.common_gui import (
+from kohya_gui.common_gui import (
     get_file_path,
     get_saveasfile_path,
     color_aug_changed,
@@ -24,27 +24,27 @@ from library.common_gui import (
     SaveConfigFile,
     save_to_file,
 )
-from library.class_configuration_file import ConfigurationFile
-from library.class_source_model import SourceModel
-from library.class_basic_training import BasicTraining
-from library.class_advanced_training import AdvancedTraining
-from library.class_folders import Folders
-from library.class_sdxl_parameters import SDXLParameters
-from library.class_command_executor import CommandExecutor
-from library.tensorboard_gui import (
+from kohya_gui.class_configuration_file import ConfigurationFile
+from kohya_gui.class_source_model import SourceModel
+from kohya_gui.class_basic_training import BasicTraining
+from kohya_gui.class_advanced_training import AdvancedTraining
+from kohya_gui.class_folders import Folders
+from kohya_gui.class_sdxl_parameters import SDXLParameters
+from kohya_gui.class_command_executor import CommandExecutor
+from kohya_gui.tensorboard_gui import (
     gradio_tensorboard,
     start_tensorboard,
     stop_tensorboard,
 )
-from library.dreambooth_folder_creation_gui import (
+from kohya_gui.dreambooth_folder_creation_gui import (
     gradio_dreambooth_folder_creation_tab,
 )
-from library.dataset_balancing_gui import gradio_dataset_balancing_tab
-from library.utilities import utilities_tab
-from library.class_sample_images import SampleImages, run_cmd_sample
+from kohya_gui.dataset_balancing_gui import gradio_dataset_balancing_tab
+from kohya_gui.utilities import utilities_tab
+from kohya_gui.class_sample_images import SampleImages, run_cmd_sample
 
-from library.custom_logging import setup_logging
-from library.localization_ext import add_javascript
+from kohya_gui.custom_logging import setup_logging
+from kohya_gui.localization_ext import add_javascript
 
 # Set up logging
 log = setup_logging()
